algo-trading/requirements.txt
pandas
pyarrow
duckdb
httpx
tenacity
pydantic
python-dateutil
PyYAML
pytest
typer
